# Optimize job details loading with empty disk cache