--- a/lib/OpenQA/Task/Job/FinalizeResults.pm
+++ b/lib/OpenQA/Task/Job/FinalizeResults.pm
@@ -18,39 +18,31 @@ use Mojo::Base 'Mojolicious::Plugin';
 
 sub register {
     my ($self, $app) = @_;
-    $app->minion->add_task(finalize_job_results => sub { _finalize_results($app, @_); });
-    $app->minion->add_task(cleanup_job_results  => sub { _cleanup_results($app, @_); });
+    $app->minion->add_task(finalize_job_results => \&_finalize_results);
 }
 
 sub _finalize_results {
-    my ($app, $minion_job, $job_id) = @_;
-    my $guard = $app->minion->guard("finalize_job_results_for_$job_id", 86400);
-    return $minion_job->finish('Blocked by another finalize job') unless $guard;
+    my ($minion_job, $openqa_job_id) = @_;
 
-    my $job    = $app->schema->resultset('Jobs')->find({id => $job_id});
-    my $errors = 0;
+    my $app = $minion_job->app;
+    return $minion_job->fail('No job ID specified.') unless defined $openqa_job_id;
+    return $minion_job->finish("A finalize_job_results job for $openqa_job_id is already active")
+      unless my $guard = $app->minion->guard("finalize_job_results_for_$openqa_job_id", 86400);
 
-    for my $module ($job->modules_with_job_prefetched) {
-        $errors++ if !$module->finalize_results;
+    # try to finalize each
+    my $openqa_job = $app->schema->resultset('Jobs')->find($openqa_job_id);
+    return $minion_job->fail("Job $openqa_job_id does not exist.") unless $openqa_job;
+    my %failed_to_finalize;
+    for my $module ($openqa_job->modules_with_job_prefetched) {
+        eval { $module->finalize_results; };
+        if (my $error = $@) { $failed_to_finalize{$module->name} = $error; }
     }
 
-    $app->gru->enqueue(cleanup_job_results => [$job_id] => {delay => 60});
-
-    if ($errors) {
-        $minion_job->fail("Finalizing results of $errors modules failed");
-    }
-}
-
-sub _cleanup_results {
-    my ($app, $minion_job, $job_id) = @_;
-    # This job must not run in parallel with _finalize_results()
-    my $guard = $app->minion->guard("finalize_job_results_for_$job_id", 86400);
-    return $minion_job->finish('Blocked by another finalize job') unless $guard;
-
-    my $job = $app->schema->resultset('Jobs')->find({id => $job_id});
-
-    for my $module ($job->modules_with_job_prefetched) {
-        $module->cleanup_results;
+    # record failed modules
+    if (%failed_to_finalize) {
+        my $count = scalar keys %failed_to_finalize;
+        $minion_job->note(failed_modules => \%failed_to_finalize);
+        $minion_job->fail("Finalizing results of $count modules failed");
     }
 }