Each snippet of console output is stored in a separate file in the job result directory. Collect these tiny files into module result JSON files at the end of each job. This can speed up server-side job details JSON generation up to 5 times when the disk cache is empty.

On my local OpenQA instance, this patch speeds up `details_ajax` JSON generation for ltp_syscalls jobs as follows:
- empty cache: from 3.3 seconds to 750ms
- full cache: from 750ms to 450ms