@mdoucha 2020-06-03
Test result cleanup reworked as minion job. I'll add a separate commit for the step view and deduplication.