@mdoucha 2020-06-04
BTW, I ran into some os-autoinst bugs yesterday while testing Unicode handling in this PR. When a command executed on SUT prints non-ASCII UTF-8 output, it'll show up correctly in `serial_terminal.txt` but `wait_serial()` will return binary garbage that was most likely passed through `encode('UTF-8', $str)` twice too many.