@mdoucha 2020-06-05
Finalize jobs will now be marked as failed if any of the modules doesn't get properly finalized. Also added a comment explaining the `enqueue()` conditions in `Jobs::store_column()`.