@Martchus 2020-06-26
@mdoucha told me he will spent no further time on this PR so I'm taking over. It would be a waste to throw the code away at this point and this is actually a beneficial optimization.