@Martchus 2020-06-29
According to the checks our criteria for test coverage is reached. However, I haven't even tested this sufficiently manually and checked whether the unit tests are sufficient.