@Martchus 2020-06-29
The existing tests covered everything except the error handling which is now covered as well. I also tested it locally on a real job and it works.