@AdamWill 2020-09-15
I think this may have [broken stale job detection](https://progress.opensuse.org/issues/71386). Generally speaking it seems dangerous to have DB schema code assuming whatever 'app' is using it will have a `$self->gru`, if I'm understanding what `OpenQA::App->singleton` does correctly?