@okurz 2020-09-16
Do you mean that the conditional execution of `$app->gru->enqueue(finalize_job_results...` does not ensure that the job ever finishes?