@Martchus 2020-09-16
https://github.com/os-autoinst/openQA/pull/3389 will fix it implicitly because I amended the condition to make tests pass. However, the solution isn't ideal. Maybe it makes sense to initialize the gru plugin within the scheduler.