@kraih 2020-09-16
Since gru is part of the model it's an easy argument to make that it should be shared between all webui micro services.