@Martchus 2020-09-16
@kraih Yes, I've created https://github.com/os-autoinst/openQA/pull/3397 for that.